+++ src/js/ajax_utils.js
@@ -1,8 +1,7 @@
-function ajaxCheckPid(base, method, pid, ahash, interval) {
+function ajaxCheckPid(base, method, pid, interval) {
     // base is a URL base, e.g. https://cmsweb.cern.ch
     // method is request method, e.g. /request
     // pid is DASQuery qhash
-    // ahash is user args hash
     // status request interval in seconds
     var limit = 30000; // in miliseconds
     var wait  = parseInt(interval);
@@ -11,7 +10,7 @@ function ajaxCheckPid(base, method, pid, ahash, interval) {
     } else { wait = limit; }
     new Ajax.Updater('response', base+'/'+method,
     { method: 'get' ,
-      parameters : {'pid': pid, 'ahash': ahash},
+      parameters : {'pid': pid},
       onException: function() {return;},
       onComplete : function() {
         if (url.indexOf('view=xml') != -1 ||
@@ -22,10 +21,11 @@ function ajaxCheckPid(base, method, pid, ahash, interval) {
         var sec = wait/1000;
         var msg = ', next check in '+sec.toString()+' sec, please wait..., <a href="/das/">stop</a> request';
         // look at transport body and match its content,
-        // if check_pid still working on request, call again
+        // if check_pid still working on request, call again, otherwise
+        // reload the request page
         if (transport.responseText.match(/processing PID/)) {
             transport.responseText += msg;
-            setTimeout('ajaxCheckPid("'+base+'","'+method+'","'+pid+'","'+ahash+'","'+wait+'")', wait);
+            setTimeout('ajaxCheckPid("'+base+'","'+method+'","'+pid+'","'+wait+'")', wait);
         } else {
             reload();
         }