+++ src/python/DAS/web/das_web_srv.py
@@ -210,7 +210,7 @@ class DASWebService(DASWebManager):
                 headers=cherrypy.request.headers,
                 method=cherrypy.request.method,
                 path=cherrypy.request.path_info,
-                args=args, ahash=genkey(args),
+                args=args,
                 ip=cherrypy.request.remote.ip,
                 hostname=cherrypy.request.remote.name,
                 port=cherrypy.request.remote.port)
@@ -761,7 +761,6 @@ class DASWebService(DASWebManager):
         uinput  = kwargs.get('input', '')
         if  self.busy():
             return self.busy_page(uinput)
-        ahash    = genkey(cherrypy.request.params)
         self.logdb(uinput)
         form    = self.form(uinput=uinput, instance=inst, view=view)
         check, content = self.generate_dasquery(uinput, inst)
@@ -786,7 +785,7 @@ class DASWebService(DASWebManager):
             self.reqmgr.add(pid, kwargs)
             if  self.taskmgr.is_alive(pid):
                 page = self.templatepage('das_check_pid', method='check_pid',
-                        uinput=uinput, view=view, ahash=ahash,
+                        uinput=uinput, view=view,
                         base=self.base, pid=pid, interval=self.interval)
             else:
                 page = self.get_page_content(kwargs)
@@ -812,12 +811,11 @@ class DASWebService(DASWebManager):
         return self.page(page)
 
     @expose
-    @checkargs(['pid', 'ahash'])
-    def check_pid(self, pid, ahash):
+    @checkargs(['pid'])
+    def check_pid(self, pid):
         """
-        Check status of given pid and return appropriate page content.
-        This is a server callback function for ajaxCheckPid, see
-        js/ajax_utils.js
+        Check status of given pid. This is a server callback
+        function for ajaxCheckPid, see js/ajax_utils.js
         """
         # do not allow caching
         set_no_cache_flags()
@@ -828,19 +826,7 @@ class DASWebService(DASWebManager):
             if  self.taskmgr.is_alive(pid):
                 page = img + " processing PID=%s" % pid
             else:
-                kwargs = self.reqmgr.get(pid)
-                if  kwargs and kwargs.has_key('dasquery'):
-                    del kwargs['dasquery']
-                # if no kwargs (another request delete it)
-                # use logging DB to look-up user request via ahash
-                if  not kwargs:
-                    spec = {'ahash':ahash}
-                    skey = [('ts', DESCENDING)]
-                    res  = [r for r in self.logcol.find(spec).sort(skey)]
-                    kwargs = res[0]['args']
-                    self.adjust_input(kwargs)
                 self.reqmgr.remove(pid)
-                page = self.get_page_content(kwargs)
         except Exception as err:
             msg = 'check_pid fails for pid=%s' % pid
             print dastimestamp('DAS WEB ERROR '), msg